% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closestPix.R
\name{closestPix}
\alias{closestPix}
\title{Find the closest pixel in a raster given a set of coordinates}
\usage{
closestPix(x, y, raster)
}
\arguments{
\item{x}{A vector of X coordinates}

\item{y}{A vector of Y coordinates}

\item{raster}{A \code{\link{raster}}. This could also be a \code{\link{stack}} or a \code{\link{brick}}.}
}
\description{
Given a set of locations (x and y coordinates), find the location of the closest pixel in a raster. This function is designed to account for locations that are outside the extent of the raster. This function is meant to be used internally.
}
