% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distRaster.R
\name{distRaster}
\alias{distRaster}
\title{Distance to geometry}
\usage{
distRaster(sp, raster)
}
\arguments{
\item{sp}{A \code{SpatialPolygons}, \code{SpatialPolygonsDataFrame}, \code{SpatialLines}, \code{SpatialLinesDataFrame}, \code{SpatialPoints} or \code{SpatialPointsDataFrame} object}

\item{raster}{A \code{\link{raster}} that includes the explanatory variables to consider for the analysis. This could also be a \code{\link{stack}} or a \code{\link{brick}}.}
}
\description{
Calculates for all pixels of a map the distance to a particular geometry. This function is essential a wrapper over the \code{\link{distance}} function of the raster package
}
