% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expertMEGR-data.R
\docType{data}
\name{expertMEGR}
\alias{expertMEGR}
\title{Expert map for pseudospecies MEGR}
\format{An object of class \code{SpatialPolygons} of length 1.}
\usage{
data(expertMEGR)
}
\description{
An artificial expert map for pseudospecies MEGR
}
\details{
A \code{SpatialPolygons} that can serve as an expert map for the sake of illustrations. This map highlights all the area where the soil water content is below 400 in the sampling regions. It was constructed from the soil water content (WatrCont) layer of \code{\link{mite.envRaster}}.

The projection used for these data preserves the fine (meter) scale, which is useful for some of the illustration presented in the vignettes.
}
\keyword{datasets}
