% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explanaMesh.R
\name{explanaMesh}
\alias{explanaMesh}
\title{Prepare data for model construction}
\usage{
explanaMesh(sp, mesh, X, verbose = TRUE)
}
\arguments{
\item{sp}{A \code{SpatialPolygons} or a \code{SpatialPolygonsDataFrame}}

\item{mesh}{An \code{inla.mesh} object}

\item{X}{A \code{\link{raster}} that includes the explanatory variables to consider for the analysis. This could also be a \code{\link{stack}} or a \code{\link{brick}}.}

\item{verbose}{Logical. Whether or not to print on the screen (five times) the number of edges that have been assigned values.}
}
\description{
Prepare and organise the data for model prediction. This function is meant to be used as a starting point for all spatial and spatiotemporal model in this package. This function also gathers explanatory variables values associated to the edges within and outside the area of interest.
}
\details{
The time it takes to run this function is directly related to the number of vertex (points) in the mesh; more specifically the ones outside the region of interest.
}
