% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapSpatial.R
\name{mapSpatial}
\alias{mapSpatial}
\title{Construct maps for spatial point process models}
\usage{
mapSpatial(spatialModel, dims, type = c("mean", "sd", "0.025quant",
  "0.5quant", "0.975quant", "mode"), sp = NULL)
}
\arguments{
\item{spatialModel}{An object of class \code{spatialPP} or of class \code{spatialPA}.}

\item{dims}{A vector of length 2 defining the number of pixels in rows and columns of the map.}

\item{type}{Either "mean", "sd", "0.025quant", "0.5quant", "0.975quant" or "mode". Defines the map to be drawn.}

\item{sp}{A spatial polygon to isolate the region of interest. If none is given, a map is drawn for the entire region covered by the mesh.}
}
\description{
Constructs mean and standard deviation maps for spatial point process models
}
