% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapSpatialPP.R
\name{mapSpatialPP}
\alias{mapSpatialPP}
\title{Construct maps for spatial point process models}
\usage{
mapSpatialPP(spatialPP, mesh, resolution, type = c("mean", "sd", "0.025quant",
  "0.5quant", "0.975quant", "mode"), sp = NULL)
}
\arguments{
\item{spatialPP}{An object of class \code{spatialPP} obtained from the \code{\link{spatialPP}} function.}

\item{mesh}{An \code{inla.mesh} object.}

\item{resolution}{A vector of length 2 defining the number of pixels in rows and columns of the map.}

\item{type}{Either "mean", "sd", "0.025quant", "0.5quant", "0.975quant" or "mode". Defines the map to be drawn.}

\item{sp}{A spatial polygon to isolate the region of interest. If none is given, a map is drawn for the entire region covered by the mesh.}
}
\description{
Constructs mean and standard deviation maps for spatial point process models
}
