% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppWeight.R
\name{ppWeight}
\alias{ppWeight}
\title{Calculate weight for the integration of the point process}
\usage{
ppWeight(sPoly, mesh)
}
\arguments{
\item{sPoly}{A \code{SpatialPolygons*} object.}

\item{mesh}{An \code{inla.mesh} object}
}
\value{
An object of class \code{ppWeight} which returns a vector of weights.
}
\description{
This function calculates the weight that should be given to the
}
\keyword{models}
