% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialMesh.R
\name{spatialMesh}
\alias{spatialMesh}
\title{Build triangulation for a spatial polygon}
\usage{
spatialMesh(sp, coords = NULL, generic = TRUE, ...)
}
\arguments{
\item{sp}{A \code{SpatialPolygons}, \code{SpatialPolygonsDataFrame}, \code{SpatialLines}, \code{SpatialLinesDataFrame}, \code{SpatialPoints} or \code{SpatialPointsDataFrame} object}

\item{coords}{A 2-columns matrix of point locations}

\item{generic}{Logical. Defines a generic mesh (see details).}

\item{\dots}{Arguments passed to \code{\link{inla.mesh.2d}}}
}
\value{
An \code{inla.mesh} object
}
\description{
Builds a triangulation for a spatial polygon where a set of coordinates (usually fictive) is included to define the size of the triangles. In short, where there are less points the triangles will be larger insurance a more reliable model estimation in areas where there are less sampling efforts.
}
\details{
Although the \code{generic} argument can build a mesh designed for any spatial domain, the parameters used to build the mesh were chosen for North America.

NOT YET FULLY TESTED
NOT YET FULLY TESTED
NOT YET FULLY TESTED
NOT YET FULLY TESTED
NOT YET FULLY TESTED
}
\seealso{
\code{\link{inla.mesh.2d}}
}
\author{
F. Guillaume Blanchet
}
\keyword{graphs}
