% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialPP.R
\name{spatialPP}
\alias{spatialPP}
\title{Spatial point process model}
\usage{
spatialPP(formula, y, X, sp, mesh, smooth = 2, prior.range = c(0.05, 0.01),
  prior.sigma = c(1, 0.01), ...)
}
\arguments{
\item{formula}{A formula that only relates the response \code{y} and some (or all) of the explanatory variables \code{X}. A paricularity of the is formula is that the response has to be defined as \code{y}.}

\item{y}{A 2-columns matrix defining the location of where a species was found.}

\item{X}{A raster \code{stack} (or \code{brick}) combining all of the explanatory variables to consider.}

\item{sp}{A \code{SpatialPolygons} or \code{SpatialPolygonsDataFrame}}

\item{mesh}{An \code{inla.mesh} object}

\item{smooth}{A single value ranging between 0 and 2 passed to \code{inla.spde2.pcmatern}. It defines the smoothness of the Matern SPDE model. Default is set at 2.}

\item{prior.range}{A vector of length 2, with (range0, Prange) specifying that P(ρ < ρ_0) = p_ρ, where ρ is the spatial range of the random field. If Prange is NA, then range0 is used as a fixed range value. Default is c(0.05, 0.01).}

\item{prior.sigma}{A vector of length 2, with (sigma0, Psigma) specifying that P(σ > σ_0) = p_σ, where σ is the marginal standard deviation of the field. If Psigma is NA, then sigma0 is used as a fixed range value.  Default is c(1, 0.01).}

\item{\dots}{Arguments passed to \code{inla}}
}
\value{
An object of class \code{\link{inla}}
}
\description{
Spatial point process model using INLA. This function is essentially a sophisticated wrapper over \code{inla}
}
\keyword{models}
