% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temporalMesh.R
\name{temporalMesh}
\alias{temporalMesh}
\title{Build a 1 dimensional mesh for a temporal model}
\usage{
temporalMesh(knots = NULL, ...)
}
\arguments{
\item{knots}{A vector defining the temporal location of the knots. Default is NULL}

\item{\dots}{Arguments passed to \code{\link{inla.mesh.1d}}}
}
\value{
An \code{inla.mesh} object
}
\description{
Builds a 1 dimensional mesh for a temporal model. This is essential a generic wrapper for inla.mesh.1d
}
\details{
If \code{knots} is \code{NULL}, the mesh is constructed for one year with a daily resolution where the knots are located at the beginning of each month and a last knot is at the end of the year.
}
\seealso{
\code{\link{inla.mesh.1d}}
}
\author{
F. Guillaume Blanchet
}
\keyword{graphs}
