% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniSpace.R
\name{uniSpace}
\alias{uniSpace}
\title{Univariate spatial model}
\usage{
uniSpace(formula, sPointsDF, explanaMesh, offset = NULL, family = NULL,
  link = NULL, smooth = 2, prior.range = c(0.05, 0.01),
  prior.sigma = c(1, 0.01), ...)
}
\arguments{
\item{formula}{A formula that relates the response and some (or all) of the explanatory variables \code{X}. The name of the response variable is the name of the variable in \code{sPointsDF}.}

\item{sPointsDF}{A SpatialPointsDataFrame with a vector of 0 and 1 in the data.frame part \code{stack} (or \code{brick}) combining all of the explanatory variables to consider.}

\item{explanaMesh}{An object of class \code{explanaMesh}}

\item{offset}{A character string defining the explanatory variable in \code{explanaMesh$X} to use as offset.}

\item{family}{A character string describing the error distribution to be used when constructing the model.}

\item{link}{A character string describing the link function to be used when constructing the model.}

\item{smooth}{A single value ranging between 0 and 2 passed to \code{inla.spde2.pcmatern}. It defines the smoothness of the Matern SPDE model. Default is set at 2.}

\item{prior.range}{A vector of length 2, with (range0, Prange) specifying that P(ρ < ρ_0) = p_ρ, where ρ is the spatial range of the random field. If Prange is NA, then range0 is used as a fixed range value. Default is c(0.05, 0.01).}

\item{prior.sigma}{A vector of length 2, with (sigma0, Psigma) specifying that P(σ > σ_0) = p_σ, where σ is the marginal standard deviation of the field. If Psigma is NA, then sigma0 is used as a fixed range value.  Default is c(1, 0.01).}

\item{\dots}{Arguments passed to \code{inla}}
}
\value{
An object of class \code{uniSpace} that includes a model output, which is the model output of INLA.

In addition, it includes a series of attributes:

  \item{\code{formula}}{The formula used to construct the model}
  \item{\code{sPointsDF}}{A \code{SpatialPointDataFrame} object that includes the sample location and associated data of the modelled species.}
  \item{\code{XEst}}{A matrix with all the explanatory variables used to construct the model. If there were factors in the original set of explanatory variables \code{X}, in \code{XEst}, they were decomposed into dummy variables. The values in \code{XEst} are the one from the sampled location.}
  \item{\code{XPred}}{A matrix with all the explanatory variables used to construct the model. If there were factors in the original set of explanatory variables \code{X}, in \code{XPred}, they were decomposed into dummy variables. The values in \code{XPred} were gathered at the mesh edges.}
  \item{\code{mesh}}{An object of class \code{inla.mesh}. It is the mesh used to construct the model.}
  \item{\code{Stack}}{An object of class \code{inla.data.stack}. It is a stack object constructed internally.}
}
\description{
Spatial model for presence-absence data using INLA. This function is essentially a specialized wrapper over \code{inla}
}
\details{
The underlying model used by this function is a generalized spatial linear model using the cloglog link function. The idea to use the cloglog link function instead of another link function (e.g. logit or probit) is that it is more flexible.
}
\keyword{models}
