% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weightPP.R
\name{weightPP}
\alias{weightPP}
\title{Calculate weight for the integration of the point process}
\usage{
weightPP(sp, mesh)
}
\arguments{
\item{sp}{A \code{SpatialPolygons} or \code{SpatialPolygonsDataFrame} object.}

\item{mesh}{An \code{inla.mesh} object}
}
\description{
Calculate weight for the integration of the point process
}
\keyword{models}
